// src/lib/generated/models/WalletTreeConfig.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type WalletTreeConfig = {
    /**
     * The currency of the virtual account
     */
    apiKey: string
    /**
     * The currency of the virtual account
     */
    ceramicUrl?: string
    /**
     * The currency of the virtual account
     */
    sessionKeyId?: string
}